// WebApp/src/app/interfaces/Resultado.interface.ts
// Generated by https://quicktype.io

import { Filter } from "./filtroValidados.interface";

export interface Resultado {
  noEntregaOCDL: string;
  claveUnica: string;
  claveSitioOriginal: string;
  claveSitio: string;
  claveMonitoreo: string;
  nombreSitio: string;
  tipoSitio: string;
  claveParametro: string;
  laboratorio: string;
  tipoCuerpoAgua: string;
  tipoCuerpoAguaOriginal: string; 
  resultado: string;
  tipoAprobacion: string;
  esCorrectoResultado: string;
  observaciones: string;
  fechaRevision: string;
  fechaLimiteRevision: string;
  nombreUsuario: string;
  estatusResultado: string;
  isChecked: boolean;
  isCheckedmodal: boolean;
  ocdl?: string;
  organismoCuenca: string;
  DireccionLocal?: string;
  fechaRealizacion: string;
  //filtrosge?: Filter[];
  direccionLocal: string;
  lstParametros?: any[];
  lstEvidencias?: any[];
  lstParametrosTotalOrden?: any[];
  lstParametrosTotalResultado?: any[];
  cuerpoAgua: string;
  cuerpoAguaId: any;
  tipoCuerpoAguaId: any;
  tipoHomologado: string;
  tipoHomologadoId: any;

  muestreoId: any;
  UsuarioId?: any;
  TipoAprobId?: any;
  lstparametros?: any[];  



  noEntrega: string;

  cuerpoAguaO?: string;
  resultadoAR?: string;
  escorrectoROCDL?: string;
  observacionesOCDL?: string;
  esCorectoresultado?: string;
  observacionSecaia?: string;
  clasificaObserva?: string;
  observacionSRE?: string;
  comentarios: string;
  fechaObservaSRE?: string;
  seaprueba?: string;
  fechaEstatus?: string;
  estatus?: string;
  estatusFinal?: string;
  //Replica-Resumen
  clasificacionObservacion?: any;
  seApruebaElResultado?: any;
  fechaAprobacionRechazo?: any;
  nombreUsuarioQueReviso?: any;
  estatusDelResultado?: any;
  esCorrectoElResultadoPorOCDL?: any;
  observacionOCDL?: string;
  esCorrectoElResultadoPorSECAIA?: any;
  observacionSECAIA?: any;

  //Replica-Total
  corcResultSECAIA: string;
  obsSECAIA: string;
  causRechazo: string;
  acepRechazo: string;
  resultRep: string;
  mismoResult: string;
  obsLab: string;
  fechaRepLab: string;
  obsSRENAMECA: string;
  apruebaResultDespRep: any;
  nomUsRevision: string;

 
}

export interface ResultadoDescarga {
  noEntregaOCDL: string;
  ocdl?: string;
  nombreSitio: string;
  claveMonitoreo: string;
  fechaRealizacion: string;
  laboratorio: string; 
  tipoCuerpoAgua: string;
  lstParametros?: any[];
  lstParametrosOrden?: any[];
}